+++ packages/orq-rc/RELEASES.md
@@ -6919,3 +6919,13 @@ Based on:
 - [typescript v4.4.0-rc.9] packages/orq-rc
 ### Releases
 - [NPM v4.4.0-rc.9] https://www.npmjs.com/package/@orq-ai/node/v/4.4.0-rc.9 - packages/orq-rc
+
+## 2026-02-18 12:48:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.722.2 (2.832.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.4.0-rc.10] packages/orq-rc
+### Releases
+- [NPM v4.4.0-rc.10] https://www.npmjs.com/package/@orq-ai/node/v/4.4.0-rc.10 - packages/orq-rc